// src/generated/schema.ts
export interface CharacterDescriptionArgs {
  /** Return the string in pre-parsed html instead of markdown */
  readonly asHtml?: boolean | null;
}

export interface CharacterMediaArgs {
  readonly sort?: Array<MediaSort | null> | null;
  readonly type?: MediaType | null;
  readonly onList?: boolean | null;
  /** The page */
  readonly page?: number | null;
  /** The amount of entries per page, max 25 */
  readonly perPage?: number | null;
}
// ...
export interface QueryPageArgs {
  /** The page number */
  readonly page?: number | null;
  /** The amount of entries per page, max 50 */
  readonly perPage?: number | null;
}
// ...
export interface QueryMediaTrendArgs {
  /** Filter by the media id */
  readonly mediaId?: number | null;
  /** Filter by date */
  readonly date?: number | null;
  /** Filter by trending amount */
  readonly trending?: number | null;
  /** Filter by score */
  readonly averageScore?: number | null;
  /** Filter by popularity */
  readonly popularity?: number | null;
  /** Filter by episode number */
  readonly episode?: number | null;
  /** Filter to stats recorded while the media was releasing */
  readonly releasing?: boolean | null;
  /** Filter by the media id */
  readonly mediaId_not?: number | null;
  /** Filter by the media id (max 10,000 items) */
  readonly mediaId_in?: Array<number | null> | null;
  /** Filter by the media id (max 10,000 items) */
  readonly mediaId_not_in?: Array<number | null> | null;
  /** Filter by date */
  readonly date_greater?: number | null;
  /** Filter by date */
  readonly date_lesser?: number | null;
  /** Filter by trending amount */
  readonly trending_greater?: number | null;
  /** Filter by trending amount */
  readonly trending_lesser?: number | null;
  /** Filter by trending amount */
  readonly trending_not?: number | null;
  /** Filter by score */
  readonly averageScore_greater?: number | null;
  /** Filter by score */
  readonly averageScore_lesser?: number | null;
  /** Filter by score */
  readonly averageScore_not?: number | null;
  /** Filter by popularity */
  readonly popularity_greater?: number | null;
  /** Filter by popularity */
  readonly popularity_lesser?: number | null;
  /** Filter by popularity */
  readonly popularity_not?: number | null;
  /** Filter by episode number */
  readonly episode_greater?: number | null;
  /** Filter by episode number */
  readonly episode_lesser?: number | null;
  /** Filter by episode number */
  readonly episode_not?: number | null;
  /** The order the results will be returned in */
  readonly sort?: Array<MediaTrendSort | null> | null;
}
// ...
export interface QueryMediaTagCollectionArgs {
  /** Mod Only */
  readonly status?: number | null;
}
// ...
export interface AiringScheduleSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly airingAt?: true;
  readonly timeUntilAiring?: true;
  readonly episode?: true;
  readonly mediaId?: true;
  readonly media?: MediaSelection;
}
// ...
export interface CharacterSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly name?: CharacterNameSelection;
  readonly image?: CharacterImageSelection;
  readonly description?: { readonly $args: CharacterDescriptionArgs; readonly $select: true };
  readonly gender?: true;
  readonly dateOfBirth?: FuzzyDateSelection;
  readonly age?: true;
  readonly bloodType?: true;
  readonly isFavourite?: true;
  readonly isFavouriteBlocked?: true;
  readonly siteUrl?: true;
  readonly media?: { readonly $args: CharacterMediaArgs; readonly $select: MediaConnectionSelection };
  readonly updatedAt?: true;
  readonly favourites?: true;
  readonly modNotes?: true;
}
// ...
export interface CharacterImageSelection {
  readonly __typename?: true;
  readonly large?: true;
  readonly medium?: true;
}

export interface CharacterNameSelection {
  readonly __typename?: true;
  readonly first?: true;
  readonly middle?: true;
  readonly last?: true;
  readonly full?: true;
  readonly native?: true;
  readonly alternative?: true;
  readonly alternativeSpoiler?: true;
  readonly userPreferred?: true;
}
// ...
export interface FuzzyDateSelection {
  readonly __typename?: true;
  readonly year?: true;
  readonly month?: true;
  readonly day?: true;
}
// ...
export interface MediaSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly idMal?: true;
  readonly title?: MediaTitleSelection;
  readonly type?: true;
  readonly format?: true;
  readonly status?: { readonly $args: MediaStatusArgs; readonly $select: true };
  readonly description?: { readonly $args: MediaDescriptionArgs; readonly $select: true };
  readonly startDate?: FuzzyDateSelection;
  readonly endDate?: FuzzyDateSelection;
  readonly season?: true;
  readonly seasonYear?: true;
  readonly seasonInt?: true;
  readonly episodes?: true;
  readonly duration?: true;
  readonly chapters?: true;
  readonly volumes?: true;
  readonly countryOfOrigin?: true;
  readonly isLicensed?: true;
  readonly source?: { readonly $args: MediaSourceArgs; readonly $select: true };
  readonly hashtag?: true;
  readonly trailer?: MediaTrailerSelection;
  readonly updatedAt?: true;
  readonly coverImage?: MediaCoverImageSelection;
  readonly bannerImage?: true;
  readonly genres?: true;
  readonly synonyms?: true;
  readonly averageScore?: true;
  readonly meanScore?: true;
  readonly popularity?: true;
  readonly isLocked?: true;
  readonly trending?: true;
  readonly favourites?: true;
  readonly tags?: MediaTagSelection;
  readonly relations?: MediaConnectionSelection;
  readonly characters?: { readonly $args: MediaCharactersArgs; readonly $select: CharacterConnectionSelection };
  readonly staff?: { readonly $args: MediaStaffArgs; readonly $select: StaffConnectionSelection };
  readonly studios?: { readonly $args: MediaStudiosArgs; readonly $select: StudioConnectionSelection };
  readonly isFavourite?: true;
  readonly isFavouriteBlocked?: true;
  readonly isAdult?: true;
  readonly nextAiringEpisode?: AiringScheduleSelection;
  readonly airingSchedule?: { readonly $args: MediaAiringScheduleArgs; readonly $select: AiringScheduleConnectionSelection };
  readonly trends?: { readonly $args: MediaTrendsArgs; readonly $select: MediaTrendConnectionSelection };
  readonly externalLinks?: MediaExternalLinkSelection;
  readonly streamingEpisodes?: MediaStreamingEpisodeSelection;
  readonly rankings?: MediaRankSelection;
  readonly mediaListEntry?: MediaListSelection;
  readonly reviews?: { readonly $args: MediaReviewsArgs; readonly $select: ReviewConnectionSelection };
  readonly recommendations?: { readonly $args: MediaRecommendationsArgs; readonly $select: RecommendationConnectionSelection };
  readonly stats?: MediaStatsSelection;
  readonly siteUrl?: true;
  readonly autoCreateForumThread?: true;
  readonly isRecommendationBlocked?: true;
  readonly isReviewBlocked?: true;
  readonly modNotes?: true;
}

export interface MediaCharacterSelection {
  readonly __typename?: true;
  readonly id?: true;
  readonly role?: true;
  readonly roleNotes?: true;
  readonly dubGroup?: true;
  readonly characterName?: true;
  readonly character?: CharacterSelection;
  readonly voiceActor?: StaffSelection;
}

export interface MediaConnectionSelection {
  readonly __typename?: true;
  readonly edges?: MediaEdgeSelection;
  readonly nodes?: MediaSelection;
  readonly pageInfo?: PageInfoSelection;
}
// ...
export interface MediaTrendSelection {
  readonly __typename?: true;
  readonly mediaId?: true;
  readonly date?: true;
  readonly trending?: true;
  readonly averageScore?: true;
  readonly popularity?: true;
  readonly inProgress?: true;
  readonly releasing?: true;
  readonly episode?: true;
  readonly media?: MediaSelection;
}
// ...
export interface PageSelection {
  readonly __typename?: true;
  readonly pageInfo?: PageInfoSelection;
  readonly users?: { readonly $args: PageUsersArgs; readonly $select: UserSelection };
  readonly media?: { readonly $args: PageMediaArgs; readonly $select: MediaSelection };
  readonly characters?: { readonly $args: PageCharactersArgs; readonly $select: CharacterSelection };
  readonly staff?: { readonly $args: PageStaffArgs; readonly $select: StaffSelection };
  readonly studios?: { readonly $args: PageStudiosArgs; readonly $select: StudioSelection };
  readonly mediaList?: { readonly $args: PageMediaListArgs; readonly $select: MediaListSelection };
  readonly airingSchedules?: { readonly $args: PageAiringSchedulesArgs; readonly $select: AiringScheduleSelection };
  readonly mediaTrends?: { readonly $args: PageMediaTrendsArgs; readonly $select: MediaTrendSelection };
  readonly notifications?: { readonly $args: PageNotificationsArgs; readonly $select: NotificationUnionSelection };
  readonly followers?: { readonly $args: PageFollowersArgs; readonly $select: UserSelection };
  readonly following?: { readonly $args: PageFollowingArgs; readonly $select: UserSelection };
  readonly activities?: { readonly $args: PageActivitiesArgs; readonly $select: ActivityUnionSelection };
  readonly activityReplies?: { readonly $args: PageActivityRepliesArgs; readonly $select: ActivityReplySelection };
  readonly threads?: { readonly $args: PageThreadsArgs; readonly $select: ThreadSelection };
  readonly threadComments?: { readonly $args: PageThreadCommentsArgs; readonly $select: ThreadCommentSelection };
  readonly reviews?: { readonly $args: PageReviewsArgs; readonly $select: ReviewSelection };
  readonly recommendations?: { readonly $args: PageRecommendationsArgs; readonly $select: RecommendationSelection };
  readonly likes?: { readonly $args: PageLikesArgs; readonly $select: UserSelection };
}

export interface PageInfoSelection {
  readonly __typename?: true;
  readonly total?: true;
  readonly perPage?: true;
  readonly currentPage?: true;
  readonly lastPage?: true;
  readonly hasNextPage?: true;
}

export interface ParsedMarkdownSelection {
  readonly __typename?: true;
  readonly html?: true;
}

export interface QuerySelection {
  readonly __typename?: true;
  readonly Page?: { readonly $args: QueryPageArgs; readonly $select: PageSelection };
  readonly Media?: { readonly $args: QueryMediaArgs; readonly $select: MediaSelection };
  readonly MediaTrend?: { readonly $args: QueryMediaTrendArgs; readonly $select: MediaTrendSelection };
  readonly AiringSchedule?: { readonly $args: QueryAiringScheduleArgs; readonly $select: AiringScheduleSelection };
  readonly Character?: { readonly $args: QueryCharacterArgs; readonly $select: CharacterSelection };
  readonly Staff?: { readonly $args: QueryStaffArgs; readonly $select: StaffSelection };
  readonly MediaList?: { readonly $args: QueryMediaListArgs; readonly $select: MediaListSelection };
  readonly MediaListCollection?: { readonly $args: QueryMediaListCollectionArgs; readonly $select: MediaListCollectionSelection };
  readonly GenreCollection?: true;
  readonly MediaTagCollection?: { readonly $args: QueryMediaTagCollectionArgs; readonly $select: MediaTagSelection };
  readonly User?: { readonly $args: QueryUserArgs; readonly $select: UserSelection };
  readonly Viewer?: UserSelection;
  readonly Notification?: { readonly $args: QueryNotificationArgs; readonly $select: NotificationUnionSelection };
  readonly Studio?: { readonly $args: QueryStudioArgs; readonly $select: StudioSelection };
  readonly Review?: { readonly $args: QueryReviewArgs; readonly $select: ReviewSelection };
  readonly Activity?: { readonly $args: QueryActivityArgs; readonly $select: ActivityUnionSelection };
  readonly ActivityReply?: { readonly $args: QueryActivityReplyArgs; readonly $select: ActivityReplySelection };
  readonly Following?: { readonly $args: QueryFollowingArgs; readonly $select: UserSelection };
  readonly Follower?: { readonly $args: QueryFollowerArgs; readonly $select: UserSelection };
  readonly Thread?: { readonly $args: QueryThreadArgs; readonly $select: ThreadSelection };
  readonly ThreadComment?: { readonly $args: QueryThreadCommentArgs; readonly $select: ThreadCommentSelection };
  readonly Recommendation?: { readonly $args: QueryRecommendationArgs; readonly $select: RecommendationSelection };
  readonly Like?: { readonly $args: QueryLikeArgs; readonly $select: UserSelection };
  readonly Markdown?: { readonly $args: QueryMarkdownArgs; readonly $select: ParsedMarkdownSelection };
  readonly AniChartUser?: AniChartUserSelection;
  readonly SiteStatistics?: SiteStatisticsSelection;
  readonly ExternalLinkSourceCollection?: { readonly $args: QueryExternalLinkSourceCollectionArgs; readonly $select: MediaExternalLinkSelection };
}